Data.AffineSpace

Description

Affine spaces.

Documentation

class AdditiveGroup (Diff p) => AffineSpace p where

Minimal complete definition

Nothing

Associated Types

type Diff p

Associated vector space

type Diff p = GenericDiff p

Methods

(.-.) :: p -> p -> Diff p infix 6

Subtract points

default (.-.) :: (Generic p, Diff p ~ GenericDiff p, AffineSpace (VRep p)) => p -> p -> Diff p

(.+^) :: p -> Diff p -> p infixl 6

Point plus vector

default (.+^) :: (Generic p, Diff p ~ GenericDiff p, AffineSpace (VRep p)) => p -> Diff p -> p

(.-^) :: AffineSpace p => p -> Diff p -> p infixl 6

Point minus vector

distanceSq :: (AffineSpace p, v ~ Diff p, InnerSpace v) => p -> p -> Scalar v

Square of the distance between two points. Sometimes useful for efficiency. See also distance.

distance :: (AffineSpace p, v ~ Diff p, InnerSpace v, s ~ Scalar v, Floating (Scalar v)) => p -> p -> s

Distance between two points. See also distanceSq.

alerp :: (AffineSpace p, VectorSpace (Diff p)) => p -> p -> Scalar (Diff p) -> p

Affine linear interpolation. Varies from p to p' as s varies from 0 to 1. See also lerp (on vector spaces).

affineCombo :: (AffineSpace p, v ~ Diff p, VectorSpace v) => p -> [(p, Scalar v)] -> p

Compute an affine combination (weighted average) of points. The first element is used as origin and is weighted such that all coefficients sum to 1. For example,

affineCombo a [(0.3,b), (0.2,c)]

is equal to

a .+^ (0.3 *^ (b .-. a) ^+^ 0.2 *^ (c .-. a))

and if a, b, and c were in a vector space would also be equal to

0.5 *^ a ^+^ 0.3 *^ b ^+^ 0.2 *^ c

See also linearCombo (on vector spaces).